+++ anyblok/tests/__init__.py
@@ -5,7 +5,7 @@
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
-from anyblok._argsparse import ArgsParseManager
+from anyblok.config import Configuration
 
 
-ArgsParseManager.load_config_for_test()
+Configuration.load_config_for_test()