+++ doc/CHANGES.rst
@@ -16,6 +16,7 @@ Future
 
     Break the compatibility with the previous version of anyblok
 
+* [REF] rename **ArgsParseManager** by **Configuration**
 * [REF] rename **reload_module_if_blok_is_reloaded** by 
         **reload_module_if_blok_is_reloading** method on blok
 * [REF] rename **import_cfg_file** by **import_file** method on blok